feat: Add CRUD methods in BookingService and update field types
- Added CRUD methods in BookingService: `createBooking`, `getPageOfBookings`, `getById`, and `deleteBooking`.
- Implemented helper functions for validating and retrieving Room and Customer entities.
- Updated field types (e.g., bookingId, roomId, customerId) from `int` to `Long` in DTOs and entities for better compatibility.
- Integrated BookingMapper for entity-to-DTO conversions.
- Added exception handling for invalid input and missing entities.

+++ src/main/java/com/example/hotel_booking_system/controllers/HotelController.java
@@ -39,14 +39,14 @@ public class HotelController {
 
     // Get a hotel by its ID
     @GetMapping("/{id}")
-    public ResponseEntity<HotelDTO> getHotelById(@PathVariable int id) {
+    public ResponseEntity<HotelDTO> getHotelById(@PathVariable Long id) {
         HotelDTO hotel = hotelService.getById(id);
         return ResponseEntity.ok(hotel);
     }
 
     // Update a hotel by its ID
     @PutMapping("/{id}")
-    public ResponseEntity<HotelDTO> updateHotelById(@PathVariable int id,@RequestBody HotelDTO hotelDTO) {
+    public ResponseEntity<HotelDTO> updateHotelById(@PathVariable Long id,@RequestBody HotelDTO hotelDTO) {
         HotelDTO updatedHotel = hotelService.updateHotelById(id, hotelDTO);
         return ResponseEntity.ok(updatedHotel);
     }
